Move isDuplicate error to the new file as well
